refactor DetailContainer
