+++ src/components/TelematicReceiptDetail/TelematicReceiptDetail.tsx
@@ -1,20 +1,15 @@
 import { Download } from '@mui/icons-material';
 import { Grid } from '@mui/material';
 import { useTranslation } from 'react-i18next';
-import DetailContainer from './DetailContainer';
 import TitleComponent from '../TitleComponent/TitleComponent';
+import DetailContainer, { DetailData } from '../DetailContainer/DetailContainer';
 
-export interface TelematicReceiptDetailData {
-  label: string;
-  value: string;
-  variant?: 'body1' | 'body2' | 'h6' | 'subtitle1' | 'monospaced';
-}
 
 export const TelematicReceiptDetail = () => {
 
   const { t } = useTranslation();
 
-  const summaryData: TelematicReceiptDetailData[] = [
+  const summaryData: DetailData[] = [
     { label: 'IUV', value: '0300330000000001', variant: 'monospaced' },
     { label: 'Importo', value: '50,00 €' },
     { label: 'Causale', value: 'TARI 2024' },
@@ -23,7 +18,7 @@ export const TelematicReceiptDetail = () => {
     { label: 'CF / Partita IVA', value: 'BNCMRA82B42C933X (Persona fisica)' },
   ];
 
-  const paymentData: TelematicReceiptDetailData[] = [
+  const paymentData: DetailData[] = [
     { label: 'Data esito', value: '01/09/2024' },
     { label: 'Versante', value: 'Paolo Rossi' },
     { label: 'CF / Partita IVA', value: 'PLRSRA82B42C933X (Persona fisica)' },
@@ -50,10 +45,14 @@ export const TelematicReceiptDetail = () => {
       />
       <Grid container spacing={3}>
         <Grid item lg={6} md={6}>
-          <DetailContainer title={t('commons.summary')} data={summaryData} />
+          <DetailContainer 
+            sections={[{title: t('commons.summary'), data: [...summaryData]}]} 
+          />
         </Grid>
         <Grid item lg={6} md={6}>
-          <DetailContainer title={t('commons.payment')} data={paymentData} />
+          <DetailContainer 
+            sections={[{title: t('commons.payment'), data: [...paymentData]}]}
+          />
         </Grid>
       </Grid>
     </>
